backend/requirements.txt
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0
fpdf==1.7.2
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
sorl-thumbnail==12.8.0